pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Simone Cittadini <simone.cittadini@protonmail.com>"]
readme = "README.md"
packages = [
    {include = "grim/dice",from = "components"},
    {include = "grim/stats",from = "components"},
    {include = "grim/character",from = "components"},
]



[tool.poetry.dependencies]
python = "^3.13"
textual = "^1.0.0"

[tool.poetry.group.dev.dependencies]
textual-dev = "^1.7.0"
ruff = "^0.8.6"
pytest-icdiff = "^0.9"
pytest-xdist = "^3.6.1"

[tool.ruff]
unsafe-fixes = true
line-length = 120
target-version = "py313"

[tool.ruff.lint.pycodestyle]
max-line-length = 120
max-doc-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
